% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{targetsketch-package}
\alias{targetsketch-package}
\title{targetsketch: An R/Shiny app to visualize \code{targets} pipelines.}
\description{
\code{targetsketch} is an package and R/Shiny app
for creating and visualizing \code{targets} pipelines.
}
\examples{
# targetsketch() # Launches the app.
}
\references{
\url{https://github.com/wlandau/targetsketch}
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
