% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{targetsketch}
\alias{targetsketch}
\title{Launch the targetsketch app}
\usage{
targetsketch(script = NULL)
}
\arguments{
\item{script}{Character of length 1, path to an existing targets script.
When \code{script} is \code{NULL} (default) then the function uses the default
starting script. If \code{script} file does not exist then function will stop
and user will get an error message.}
}
\description{
Launch an interactive web application for
creating and visualizing \code{targets} pipelines.
}
\examples{
\dontrun{
targetsketch()
}
}
