% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{targetsketch}
\alias{targetsketch}
\title{Launch the targetsketch app}
\usage{
targetsketch()
}
\description{
Launch an interactive web application for
creating and visualizing \code{targets} pipelines.
}
\examples{
\dontrun{
targetsketch()
}
}
